requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
weasyprint
jinja2
